+++ zerog/__init__.py
@@ -1,5 +1,7 @@
 from zerog.datastores import CouchbaseDatastore
-from zerog.handlers import BaseHandler, ProgressHandler, RunJobHandler
+from zerog.handlers import (
+    BaseHandler, GetDataHandler, ProgressHandler, RunJobHandler
+)
 from zerog.jobs import BaseJob, BaseJobSchema, NO_RESULT
 from zerog.queues import BeanstalkdQueue
 from zerog.registry import find_subclasses, import_submodules